feat:
In the code, three buttons were added: prevMusic, PauseMusic, and Next Music; with these buttons, we can control one of the functions of the media player.
Feat: We also added a volume slider, which, with the JS code, will allow us to alter the volume of the music.
Feat: with the track feature, we will be able to know and see what music we are listening