updating img functionality.
The idea is that whenever I visit the following music, the photo changes according to the singer. In the situation where we don't have a singer's photo, the music stays with a default photo. To make this function work, I utilized if and else statements. I had to create a constant called `albumPhoto` that holds the pictures. If the playlist music has a photo, it will be shown to the user; otherwise, it will display a default photo.