implement and fixing.
DarkMode is something that the user can toggle between dark and light. To accomplish this, I had to add an `onclick` toggleTheme, which makes it clickable as a button.

The toggle button wasn't working, so I had to figure out the issue. The problem was that the classes and ID were not correctly defined, so I had to correct them. I did it by writing the correct  class and ID, which I will now be able to work with as an Object-oriented